breaking App into multiple components for organization and adding API for Contact Us section
