+++ .gitignore
@@ -1,4 +1,8 @@
+*.log
+.history
+.pnp*.cjs
+.pnp*.mjs
+.yarn
+.yarnrc.yml
 build
 node_modules
-.history
-*.log
