+++ package.json
@@ -12,7 +12,7 @@
     ]
   },
   "dependencies": {
-    "@abds/react-bindings": "beta",
+    "@abds/react-bindings": "latest",
     "autoprefixer": "^10.0.2",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
@@ -33,6 +33,7 @@
   "private": true,
   "scripts": {
     "build": "react-scripts build",
+    "clean:deps": "yarn cache clean && rm -rf node_modules",
     "eject": "react-scripts eject",
     "start": "GENERATE_SOURCEMAP=false react-scripts start",
     "test": "react-scripts test"
