+++ src/components/index.js
@@ -1,0 +1,5 @@
+export * from './contactUs';
+export * from './dietForm';
+export * from './modal';
+export * from './shoppingList';
+export * from './togglePlayground';
